fixed Symfony 5
