+++ .travis.yml
@@ -8,7 +8,7 @@ cache:
 env:
   global:
     - SYMFONY_PHPUNIT_DIR="$HOME/symfony-bridge/.phpunit"
-    - SYMFONY_PHPUNIT_VERSION=6.5
+    - SYMFONY_PHPUNIT_VERSION=8.5
     - SYMFONY_DEPRECATIONS_HELPER="max[self]=0"
 
 matrix:
@@ -16,12 +16,13 @@ matrix:
   include:
     # Minimum supported dependencies with the latest and oldest PHP version
     - php: 7.4
-      env: COMPOSER_FLAGS="--prefer-lowest" SYMFONY_DEPRECATIONS_HELPER="weak"
+      env: COMPOSER_FLAGS="--prefer-lowest" SYMFONY_DEPRECATIONS_HELPER="weak" SYMFONY_PHPUNIT_VERSION=6.5
     - php: 7.1
-      env: COMPOSER_FLAGS="--prefer-lowest" SYMFONY_DEPRECATIONS_HELPER="weak"
+      env: COMPOSER_FLAGS="--prefer-lowest" SYMFONY_DEPRECATIONS_HELPER="weak" SYMFONY_PHPUNIT_VERSION=6.5
 
       # Test the latest stable release
     - php: 7.1
+      env: SYMFONY_PHPUNIT_VERSION=6.5
     - php: 7.2
     - php: 7.3
     - php: 7.4
@@ -29,9 +30,7 @@ matrix:
 
     # Test LTS versions.
     - php: 7.4
-      env: SYMFONY_REQUIRE=3.4.*
-    - php: 7.4
-      env: SYMFONY_REQUIRE=4.3.*
+      env: SYMFONY_REQUIRE=3.4.* SYMFONY_PHPUNIT_VERSION=7.5
     - php: 7.4
       env: SYMFONY_REQUIRE=4.4.*
     - php: 7.4
