+++ tests/Fixtures/App/Controller/TestController.php
@@ -2,17 +2,25 @@
 
 namespace Doctrine\Bundle\PHPCRBundle\Tests\Fixtures\App\Controller;
 
+use Doctrine\Bundle\PHPCRBundle\ManagerRegistry;
 use Doctrine\Bundle\PHPCRBundle\Tests\Fixtures\App\Document\ReferrerDocument;
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
-class TestController extends Controller
+class TestController extends AbstractController
 {
+    private $registry;
+
+    public function __construct(ManagerRegistry $registry)
+    {
+        $this->registry = $registry;
+    }
+
     /**
      * Do a simple PHPCR request.
      */
     public function phpcrRequest()
     {
-        $dm = $this->get('doctrine_phpcr')->getManager();
+        $dm = $this->registry->getManager();
 
         $document = $dm->find(null, '/foo');
