+++ tests/Fixtures/App/config/bundles.php
@@ -4,7 +4,6 @@ return [
     \Doctrine\Bundle\DoctrineBundle\DoctrineBundle::class => ['all' => true],
     \Doctrine\Bundle\PHPCRBundle\DoctrinePHPCRBundle::class => ['all' => true],
     \Symfony\Bundle\FrameworkBundle\FrameworkBundle::class => ['all' => true],
-    \Symfony\Bundle\FrameworkBundle\FrameworkBundle::class => ['all' => true],
     \Symfony\Bundle\MonologBundle\MonologBundle::class => ['all' => true],
     \Symfony\Bundle\TwigBundle\TwigBundle::class => ['all' => true],
     \Symfony\Bundle\WebProfilerBundle\WebProfilerBundle::class => ['all' => true],
