+++ tests/Fixtures/App/config/config.php
@@ -2,5 +2,7 @@
 
 $loader->import(__DIR__.'/framework.yaml');
 $loader->import(__DIR__.'/doctrine.yaml');
+$loader->import(__DIR__.'/twig.yaml');
+$loader->import(__DIR__.'/services.yaml');
 
 $loader->import(__DIR__.'/web_profiler.yaml');
