+++ tests/Fixtures/App/config/twig.yaml
@@ -1,0 +1,2 @@
+twig:
+  default_path: '%kernel.project_dir%/templates'
