+++ tests/Unit/Form/DataTransformer/PHPCRNodeToPathTransformerTest.php
@@ -5,26 +5,27 @@ namespace Doctrine\Bundle\PHPCRBundle\Tests\Unit\Form\DataTransformer;
 use Doctrine\Bundle\PHPCRBundle\Form\DataTransformer\PHPCRNodeToPathTransformer;
 use Jackalope\Node;
 use PHPCR\SessionInterface;
+use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 
 class PHPCRNodeToPathTransformerTest extends Testcase
 {
     /**
-     * @var SessionInterface|\PHPUnit_Framework_MockObject_MockObject
+     * @var SessionInterface|MockObject
      */
     private $session;
 
     /**
-     * @var PHPCRNodeToPathTransformer|\PHPUnit_Framework_MockObject_MockObject
+     * @var PHPCRNodeToPathTransformer|MockObject
      */
     private $transformer;
 
     /**
-     * @var Node|\PHPUnit_Framework_MockObject_MockObject
+     * @var Node|MockObject
      */
     private $node;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->session = $this->createMock(SessionInterface::class);
         $this->transformer = new PHPCRNodeToPathTransformer($this->session);
